refactor(application): reorganize gateway packages and update import references
- Remove repository sub-package, move gateways to domain-specific packages
- Move order gateways to gateway/order
- Move inventory gateways to gateway/inventory
- Move party gateways to gateway/party
- Move product gateways to gateway/product
- Move remaining gateways to gateway root
- Add ContactGateway
- Update all services and use cases to reflect new package structure
