+++ core/application/src/main/java/com/arka/gateway/inventory/WarehouseInventoryGateway.java
@@ -1,4 +1,4 @@
-package com.arka.gateway.repository.inventory;
+package com.arka.gateway.inventory;
 
 import com.arka.model.inventory.WarehouseInventory;
