+++ core/application/src/main/java/com/arka/gateway/order/OrderGateway.java
@@ -1,4 +1,4 @@
-package com.arka.gateway.repository.order;
+package com.arka.gateway.order;
 
 import com.arka.model.order.Order;
