+++ core/application/src/main/java/com/arka/gateway/party/SupplierGateway.java
@@ -1,4 +1,4 @@
-package com.arka.gateway.repository.party;
+package com.arka.gateway.party;
 
 import com.arka.model.Company;
