+++ core/application/src/main/java/com/arka/gateway/product/ProductGateway.java
@@ -1,16 +1,15 @@
-package com.arka.gateway.repository.product;
+package com.arka.gateway.product;
 
 import com.arka.dto.out.ProductSalesReportOut;
 import com.arka.model.product.Product;
 
 import java.time.Instant;
-import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 
 public interface ProductGateway {
 
-    Product createProduct(Product product);
+    Product create(Product product);
 
     Optional<Product> findById(Long id);
