+++ core/application/src/main/java/com/arka/service/CompanyService.java
@@ -1,7 +1,7 @@
 package com.arka.service;
 
 import com.arka.exceptions.NotFoundException;
-import com.arka.gateway.repository.party.CompanyGateway;
+import com.arka.gateway.party.CompanyGateway;
 import com.arka.model.Company;
 import lombok.RequiredArgsConstructor;
