+++ core/application/src/main/java/com/arka/service/CustomerService.java
@@ -2,8 +2,7 @@ package com.arka.service;
 
 import com.arka.dto.out.CustomerSalesReportOut;
 import com.arka.dto.value.InstantDateRange;
-import com.arka.gateway.repository.order.OrderGateway;
-import com.arka.gateway.repository.party.CustomerGateway;
+import com.arka.gateway.party.CustomerGateway;
 import com.arka.service.util.DateRangeTypeParser;
 import com.arka.service.util.DateRangeValidator;
 import lombok.RequiredArgsConstructor;
