+++ core/application/src/main/java/com/arka/service/OrderService.java
@@ -2,15 +2,13 @@ package com.arka.service;
 
 import com.arka.dto.value.InstantDateRange;
 import com.arka.exceptions.NotFoundException;
-import com.arka.gateway.repository.order.OrderGateway;
+import com.arka.gateway.order.OrderGateway;
 import com.arka.model.order.Order;
 import com.arka.service.util.DateRangeTypeParser;
 import com.arka.service.util.DateRangeValidator;
 import lombok.RequiredArgsConstructor;
-import org.springframework.cglib.core.Local;
 
 import java.math.BigDecimal;
-import java.time.Instant;
 import java.time.LocalDate;
 
 @RequiredArgsConstructor
