+++ core/application/src/main/java/com/arka/service/ProductCategoryService.java
@@ -1,7 +1,7 @@
 package com.arka.service;
 
 import com.arka.exceptions.NotFoundException;
-import com.arka.gateway.repository.product.ProductCategoryGateway;
+import com.arka.gateway.product.ProductCategoryGateway;
 import com.arka.model.product.ProductCategory;
 import lombok.RequiredArgsConstructor;
