+++ core/application/src/main/java/com/arka/service/ProductService.java
@@ -3,9 +3,7 @@ package com.arka.service;
 import com.arka.dto.out.ProductSalesReportOut;
 import com.arka.dto.value.InstantDateRange;
 import com.arka.exceptions.NotFoundException;
-import com.arka.mapper.ProductMapper;
-import com.arka.mapper.ProductMapperImpl;
-import com.arka.gateway.repository.product.ProductGateway;
+import com.arka.gateway.product.ProductGateway;
 import com.arka.model.product.Product;
 import com.arka.service.util.DateRangeTypeParser;
 import com.arka.service.util.DateRangeValidator;
