+++ core/application/src/main/java/com/arka/service/WarehouseService.java
@@ -1,7 +1,7 @@
 package com.arka.service;
 
 import com.arka.exceptions.NotFoundException;
-import com.arka.gateway.repository.WarehouseGateway;
+import com.arka.gateway.WarehouseGateway;
 import com.arka.model.Warehouse;
 import lombok.RequiredArgsConstructor;
