+++ core/application/src/main/java/com/arka/usecase/party/CreateSupplierUseCase.java
@@ -3,7 +3,7 @@ package com.arka.usecase.party;
 import com.arka.dto.in.CreateSupplierIn;
 
 import com.arka.dto.out.CompanyOut;
-import com.arka.gateway.repository.party.SupplierGateway;
+import com.arka.gateway.party.SupplierGateway;
 import com.arka.mapper.CompanyMapper;
 import com.arka.mapper.CompanyMapperImpl;
 import com.arka.model.Company;
