+++ core/application/src/main/java/com/arka/usecase/product/CreateProductUseCase.java
@@ -2,7 +2,7 @@ package com.arka.usecase.product;
 
 import com.arka.dto.in.CreateProductIn;
 import com.arka.dto.out.CreateProductOut;
-import com.arka.gateway.repository.product.ProductGateway;
+import com.arka.gateway.product.ProductGateway;
 import com.arka.mapper.CreateProductOutMapper;
 import com.arka.mapper.CreateProductOutMapperImpl;
 import com.arka.model.product.Product;
